fix : gnb 에서 분리하고 rsc payload 직렬화 우회
