+++ src/providers/MemberProvider.tsx
@@ -1,17 +1,56 @@
 "use client";
 
+import axios from "axios";
 import { useEffect } from "react";
 import { useAuthStore } from "@/store/useAuthStore";
 import { MemberProviderProps } from "@/types";
 import type { User } from "@/types";
+import { getUser } from "@/api/client/user";
+import { logoutUser } from "@/api/client/auth";
 
 interface Props extends MemberProviderProps {
   initialUser?: Pick<User, "id" | "name" | "image"> | null;
 }
 
 export function MemberProvider({ children, initialUser }: Props) {
-  // const initialized = useRef(false);
   useEffect(() => {
+    let cancelled = false;
+
+    if (typeof initialUser === "undefined") {
+      const bootstrapAuth = async () => {
+        try {
+          const me = await getUser();
+          if (cancelled) return;
+          useAuthStore.getState().setUser(me);
+        } catch (error) {
+          if (cancelled) return;
+
+          // 인증 실패는 status(401) 기준으로만 세션 정리
+          // 네트워크 오류/일시 5xx에서는 기존 세션 상태를 유지합니다.
+          const status = axios.isAxiosError(error)
+            ? error.response?.status
+            : undefined;
+
+          if (status === 401) {
+            await logoutUser().catch(() => {
+              return;
+            });
+            if (cancelled) return;
+            useAuthStore.getState().clearAuth();
+          }
+        } finally {
+          if (cancelled) return;
+          // setUser/clearAuth에서 false로 내려주지만, 예외 경로 안전망으로 한 번 더 보장
+          useAuthStore.getState().setAuthLoading(false);
+        }
+      };
+
+      void bootstrapAuth();
+      return () => {
+        cancelled = true;
+      };
+    }
+
     useAuthStore.setState((prev) => {
       if (!initialUser) {
         return {
@@ -39,6 +78,10 @@ export function MemberProvider({ children, initialUser }: Props) {
         isAuthLoading: false,
       };
     });
+
+    return () => {
+      cancelled = true;
+    };
   }, [initialUser?.id, initialUser?.name, initialUser?.image]);
 
   return <>{children}</>;
